feat: Adicionei a parte 2 da tarefa do Ely, porem ainda precisa de algumas melhores que farei logo

+++ POO - Periodo2/Atividade_Exercicio_POO_7/Questao4/src/entities/Cliente.java
@@ -82,8 +82,8 @@ public class Cliente {
         return null;
     }
 
-    public float totalizarSaldoCliente() {
-        float valorTotal = 0;
+    public double totalizarSaldoCliente() {
+        double valorTotal = 0;
         for (Conta contaCliente : this.contas) {
             valorTotal += contaCliente.consultarSaldo();
         }
