+++ POO - Periodo2/Atividade_Exercicio_POO_7/Questao4/src/entities/Poupanca.java
@@ -2,12 +2,12 @@ package entities;
 
 public class Poupanca extends Conta {
 
-    public Poupanca(String numero, float saldo, int id) {
+    public Poupanca(String numero, double saldo, int id) {
         super(numero, saldo, id);
     }
 
     public void renderJuros(){
-        float juros = getSaldo() * 0.01f;
+        double juros = getSaldo() * 0.01f;
         depositar(juros);
         System.out.println("Juros de R$" + juros + " aplicados à poupança " + getNumero());
     }
